app/config/settings.py
import os
from dataclasses import dataclass
from dotenv import load_dotenv

load_dotenv()

@dataclass(frozen=True)
class Settings:
    openai_api_key: str = os.getenv("OPENAI_API_KEY", "")
    mongodb_uri: str = os.getenv("MONGODB_URI", "")
    chroma_persist_dir: str = os.getenv("CHROMA_PERSIST_DIR", ".chroma")
    embedding_model: str = os.getenv("EMBEDDING_MODEL", "text-embedding-3-small")
    chat_model: str = os.getenv("CHAT_MODEL", "gpt-4o-mini")

settings = Settings()
