requirements.txt
streamlit
plotly
openai>=1.30.0
chromadb
pandas
numpy
PyPDF2
streamlit-option-menu
pymongo
json2table
python-dotenv
